Primer commit de Master-Opcion_2

+++ tp2_1_2.c
@@ -9,8 +9,8 @@
 int main()
 {
    int i;
-   double *p;
-   double vt[N];
+   int *p;
+   int vt[N];
    p = vt;
    srand((int)time(NULL));
